ultimos edits

+++ src/app/components/trips/detail-trip/detail-trip.component.ts
@@ -9,7 +9,7 @@ import { DomSanitizer } from '@angular/platform-browser';
 import { FormControl, FormGroup } from '@angular/forms';
 import { UsersService } from 'src/app/services/users.service';
 import { User } from 'src/app/interfaces/user.interface';
-
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-detail-trip',
@@ -73,27 +73,26 @@ export class DetailTripComponent implements OnInit {
       this.tripId = parseInt(params['tripId'])
 
       this.detail = await this.tripsService.getTripById(this.tripId);
+
       this.detail.lat = parseFloat(String(this.detail.lat));
       this.detail.lng = parseFloat(String(this.detail.lng));
 
       this.itinerary = await this.tripsService.getItineraryByTripId(this.tripId);
-      this.subscribedUsers = await this.tripsService.getSubscribedByTrip(this.tripId);
+      this.subscribedUsers = await this.tripsService.getOnlyPendingByTrip(this.tripId);
+      this.AcceptedUsers = await this.tripsService.getUsersAccepted(this.tripId);
+
+      console.log(this.AcceptedUsers);
+      console.log(this.subscribedUsers);
+
+
 
       this.userCreatorId = this.detail.user_id;
       const userData = this.usersService.getUserData();
       this.userLoggedId = userData.user_id;
       const userStatus = await this.tripsService.getUserSubscribed();
 
-      this.AcceptedUsers = await this.tripsService.getUsersAccepted(this.tripId);
-
-
       this.userCreatorName = this.detail.username;
       this.imageUserCreator = this.detail.img_user;
-
-      this.numberOfRequests = this.subscribedUsers.length;
-
-      this.AcceptedUsers = await this.tripsService.getUsersAccepted(this.tripId);
-
 
       this.numberOfRequests = this.subscribedUsers.length;
     });
@@ -124,18 +123,29 @@ export class DetailTripComponent implements OnInit {
 
   async onSubscribe() {
     const response = await this.tripsService.subscribeToTrip(this.tripId);
-    console.log(response);
+    Swal.fire({
+      icon: 'success',
+      title: '¡Te has apuntado al viaje!',
+      text: 'Espera la confirmación del organizador',
+      confirmButtonColor: '#2E8682',
+    })
+    this.subscribedUsers = await this.tripsService.getOnlyPendingByTrip(this.tripId);
+    this.AcceptedUsers = await this.tripsService.getUsersAccepted(this.tripId);
   }
 
 
   async changeStatus(aceptada: boolean, user: User) {
+
     let status = 'rechazada';
     if (aceptada) {
       status = 'aceptada';
     }
-
-
     const response = await this.tripsService.manageUsers(this.tripId, user.id, status);
+    this.subscribedUsers = await this.tripsService.getOnlyPendingByTrip(this.tripId);
+    this.AcceptedUsers = await this.tripsService.getUsersAccepted(this.tripId);
+
+    // console.log(this.subscribedUsers, this.acceptedUsers);
+
   }
 
   loadAutocomplete() {
@@ -148,13 +158,30 @@ export class DetailTripComponent implements OnInit {
     });
   }
 
-  async onDelete() {
-    const response = await this.tripsService.deleteTripById(this.tripId);
+  onDelete() {
+    Swal.fire({
+      icon: 'warning',
+      title: 'Vas a eliminar este viaje',
+      text: '¿Estás seguro?',
+      confirmButtonColor: '#2E8682',
+      showCancelButton: true,
+      cancelButtonColor: '#d33', confirmButtonText: 'Sí, lo estoy'
+    }).then(async (result) => {
+      if (result.isConfirmed) {
+        const response = await this.tripsService.deleteTripById(this.tripId);
+        await Swal.fire(
+          'Borrado',
+          'Tu viaje ha sido borrado 😟',
+          'success'
+        );
+        this.router.navigate(['/trips']);
+      }
+    })
   }
 
-  viewProfile() {
-    //this.router.navigate(['/users/'${userId}])
-
+  userLoggedIsSubscribed() {
+    return this.subscribedUsers.find((user: User) => user.id === this.userLoggedId);
 
   }
+
 }
